+++ herd-code/herd-sdk/src/tsconfig/node/build.sh
@@ -1,8 +1,15 @@
-# Clean up previous distributions
-rm -rf dist
-rm -rf build
+if [ ! -d "dist" ]; then
+  echo "cleaning dist"
+  rm -rf dist
+fi
+
+if [ ! -d "build" ]; then
+  echo "cleaning build"
+  rm -rf build
+fi
 
 if [ ! -d "node_modules" ]; then
+  echo "installing node modules"
   npm install
 fi
 
@@ -10,10 +17,14 @@ fi
 TSC="node_modules/.bin/ngc"
 ROLLUP="node_modules/.bin/rollup"
 
+echo "Running TSC on tsconfig-build.json"
 $TSC -p tsconfig-build.json
+echo "Running creating ROLLUP bundle dist/api.js"
 $ROLLUP build/api.js --output.file dist/api.js --output.format es
 
+echo "Running TSC on tsconfig-build.es5.json"
 $TSC -p tsconfig-build.es5.json
+echo "Running creating ROLLUP bundle dist/api.es5.js"
 $ROLLUP build/api.js --output.file dist/api.es5.js --output.format es
 
 rsync -a --exclude=*.js build/ dist
